backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
boto3==1.34.0
python-decouple==3.8
django-ratelimit==4.1.0
bleach==6.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1